Fix formspree.io contact form redirect
On form submitting, the formspree.io contact form seems to redirect
to their own webpage with ReCaptcha.

It seems we need to set the response datatype to JSON in the AJAX
call to avoid running into the same-origin policy issue. Set dataType
to 'json' seems to bypass this redirection and send the email as expected.